General code improvements
- Change target C# lang version to latest (and use some of those features)
- Avoid multiple enumeration
- Avoid some code repetition
- Use `DistanceToSquared` rather than `DistanceTo` where applicable
- Kleptomaniac pawns will now wait at least 850 ticks between multiple theft attempts (this prevents rapid-fire theft attempts)
- Fix big-boned pawns being able to damage thrones
- Fix martial artist pawns not being able to disarm opponents when attacking diagonally

+++ 1.6/Source/VanillaTraitsExpanded/HarmonyPatches/PawnObserver_Patches.cs
@@ -16,7 +16,7 @@ namespace VanillaTraitsExpanded
 				return;
 			}
 			Map map = ___pawn.Map;
-			for (int i = 0; (float)i < 100f; i++)
+			for (int i = 0; i < 100; i++)
 			{
 				IntVec3 intVec = ___pawn.Position + GenRadial.RadialPattern[i];
 				if (!intVec.InBounds(map) || !GenSight.LineOfSight(intVec, ___pawn.Position, map, skipFirstCell: true))
